+++ src/Octoshift/Commands/GrantMigratorRoleCommandBase.cs
@@ -3,6 +3,7 @@ using System.CommandLine;
 using System.Threading.Tasks;
 using OctoshiftCLI.Contracts;
 using OctoshiftCLI.Extensions;
+using OctoshiftCLI.Handlers;
 
 namespace OctoshiftCLI.Commands;
