+++ src/Octoshift/Handlers/CreateTeamCommandBaseHandler.cs
@@ -1,9 +1,10 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using OctoshiftCLI.Commands;
 using OctoshiftCLI.Contracts;
 
-namespace OctoshiftCLI.Commands;
+namespace OctoshiftCLI.Handlers;
 
 public class CreateTeamCommandBaseHandler
 {
