+++ src/Octoshift/Handlers/ReclaimMannequinCommandBaseHandler.cs
@@ -3,9 +3,10 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Octoshift;
+using OctoshiftCLI.Commands;
 using OctoshiftCLI.Contracts;
 
-namespace OctoshiftCLI.Commands;
+namespace OctoshiftCLI.Handlers;
 
 public class ReclaimMannequinCommandBaseHandler
 {
