+++ src/OctoshiftCLI.Tests/Octoshift/Commands/GenerateMannequinCsvCommandBaseTests.cs
@@ -5,6 +5,7 @@ using FluentAssertions;
 using Moq;
 using OctoshiftCLI.Commands;
 using OctoshiftCLI.Contracts;
+using OctoshiftCLI.Handlers;
 using OctoshiftCLI.Models;
 using Xunit;
