+++ src/OctoshiftCLI.Tests/Octoshift/Commands/ReclaimMannequinCommandBaseTests.cs
@@ -5,6 +5,7 @@ using Moq;
 using Octoshift;
 using OctoshiftCLI.Commands;
 using OctoshiftCLI.Contracts;
+using OctoshiftCLI.Handlers;
 using Xunit;
 
 namespace OctoshiftCLI.Tests.Octoshift.Commands;
