+++ src/OctoshiftCLI.Tests/Octoshift/Commands/WaitForMigrationCommandBaseTests.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using Moq;
 using OctoshiftCLI.Commands;
 using OctoshiftCLI.Contracts;
+using OctoshiftCLI.Handlers;
 using Xunit;
 
 namespace OctoshiftCLI.Tests.Octoshift.Commands;
